server/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web search API
tavily-python==0.7.7

# Logging
loguru==0.7.2

# Environment variables
python-dotenv==1.0.0

# HTTP client for external requests
httpx==0.25.2
aiohttp==3.9.1

# Additional utilities
python-multipart==0.0.6
